+++ functions/package.json
@@ -1,16 +1,14 @@
 {
   "name": "functions",
   "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "emulator": "npm run build && firebase emulators:start --project pokeapi-215911",
-    "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase functions:shell",
+    "emulator": "firebase emulators:start --project pokeapi-215911",
+    "serve": "firebase serve --only functions",
+    "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
-  "main": "lib/index.js",
+  "main": "src/index.js",
   "engines": {
     "node": "12"
   },
@@ -21,12 +19,11 @@
     "express": "^4.17.1",
     "firebase-admin": "~9.5.0",
     "firebase-functions": "^3.13.2",
-    "http-status-codes": "^2.1.4"
+    "http-status-codes": "^2.1.4",
+    "retry-axios": "^2.4.0"
   },
   "devDependencies": {
-    "firebase-tools": "^9.5.0",
-    "tslint": "~6.1.3",
-    "typescript": "~4.2.2"
+    "firebase-tools": "^9.7.0"
   },
   "private": true
 }
